moved two-filter detection cut
moved the cut that requires a detection in both filters. it was in
clean_master_catalog, and now it's in clean_point_source_catalog. I think this
makes more sense. Now in clean_master_catalog, all you have to do is match the
star identifiers.

This does have the drawback of not keeping point sources that otherwise meet
the criteria.