fixed bug in recompute_master_catalog
there was a bug caused by the fact that in principle the point source catalog
can have stars that are not present in the master catalog (i.e. if a point
source passes all the cuts but the star associated with that point source only
has detections in a single filter). fix this bug by adding a filter in
recompute_master_catalog that pulls out only the stars that are present in both
catalogs