had to restructure klip subtraction
the bookkeeping was a mess! it assumed that all the stamps for a star used the
same reference stamps, which is certainly not the case if the different target
stamps are taken in different filters. I split up the KLIP function so that each
stamp is subtracted separately instead of together with the same KL modes