new data format for subtraction results and analysis
made a new star attribute called star.pca_results, which is a dataframe that is
indexed by Kklip. This is much more convenient to work with than the previous
format where all the KLIP and subtraction results were squished into the
star.results entries.

I want to switch over to using this dataframe to store the KLIP and
Kklip-dependent results for everything, including the dashboard plots, but it
will take some time.