you can now pass a position to the pca_bias for faster execution
this will be helpful for contrast curves where we compute the PCA bias many
times for many values of Kklip. This should be faster than computing the
correlation function between the matched filters and the modes. The cost is that
there's extra overhead from shifting the MF to the correct position, but you
only have to do that once.